pyproject.toml
[project]
name = "mcp-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "fastmcp>=2.11.3",
    "langchain-mcp-adapters>=0.1.9",
    "python-dotenv>=1.1.1",
]
